requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.15.0
keras==2.15.0
mediapipe==0.10.8
pillow==10.1.0
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0